% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function Mapping Accidents by state per year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{The year that is provided}

\item{stat.num}{The number representing the state in question as defined in the
file in question.}
}
\value{
None
}
\description{
This function displays the state map with the accident location by year.
If the state number is invalid, an error is displayed.
}
\references{
2014 FARS/NASS GES Coding and Validation Manual
}
