% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_function.R
\name{fars_read}
\alias{fars_read}
\title{A File Reading Function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file being read into R}
}
\value{
This function reads in the file in a tbl_df format if the filename exists
otherwise it returns the message 'file name does not exist'
}
\description{
This function is designed to read in a file into R.
It first off checks if the file exist.
If it does not, the function stops and returns a message 'file {filename} does
not exist'.
If the file exists, it reads it into R and places it in a dataframe format
that is easily worked with.
}
\examples{
fars_read(accident_2003.csv)


}
