% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fars Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vectors with a list of years}
}
\value{
A dataframe with the months and years and provides NULL
if the years are invalid
}
\description{
This function is an ancillary function for the function fars_summarize_years
}
\examples{
fars_read_years(c(2015,2016,2017))



}
