% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{A vector with the list of years that are to be summarized}
}
\value{
Returns a summary of the number of accidents by year.
}
\description{
This function summarises the yearly accident data by month
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
