% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_function.R
\name{make_filename}
\alias{make_filename}
\title{A Make File Name Function
This function is designed to make the name of the file that is read in to
R based on the year specified.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer of the year}
}
\value{
The function returns a string with the data path and the year specified.
}
\description{
A Make File Name Function
This function is designed to make the name of the file that is read in to
R based on the year specified.
}
\examples{
make_filename(2014)

}
